[part 6] refactor!(telemetry): remove dependency on `TracerTelemetry` (#204)
Moves the TracerTelemetry logic directly into Telemetry, eliminating
the need for a separate `TracerTelemetry` object.
