+++ BUILD.bazel
@@ -43,7 +43,6 @@ cc_library(
       "src/datadog/tags.cpp",
       "src/datadog/threaded_event_scheduler.cpp",
       "src/datadog/tracer_config.cpp",
-      "src/datadog/tracer_telemetry.cpp",
       "src/datadog/tracer.cpp",
       "src/datadog/trace_id.cpp",
       "src/datadog/trace_sampler_config.cpp",
@@ -76,7 +75,6 @@ cc_library(
       "src/datadog/tag_propagation.h",
       "src/datadog/tags.h",
       "src/datadog/threaded_event_scheduler.h",
-      "src/datadog/tracer_telemetry.h",
       "src/datadog/trace_sampler.h",
       "src/datadog/w3c_propagation.h",
     ],
